Serde in separate submodules
