+++ py-polars/src/dataframe/serde.rs
@@ -1,0 +1,73 @@
+use std::io::{BufWriter, Cursor};
+use std::ops::Deref;
+
+use polars_io::mmap::ReaderBytes;
+use pyo3::prelude::*;
+use pyo3::types::PyBytes;
+
+use super::PyDataFrame;
+use crate::error::PyPolarsErr;
+use crate::file::{get_file_like, get_mmap_bytes_reader};
+use crate::prelude::*;
+
+#[pymethods]
+impl PyDataFrame {
+    #[cfg(feature = "ipc_streaming")]
+    fn __getstate__(&self, py: Python) -> PyResult<PyObject> {
+        // Used in pickle/pickling
+        let mut buf: Vec<u8> = vec![];
+        IpcStreamWriter::new(&mut buf)
+            .with_pl_flavor(true)
+            .finish(&mut self.df.clone())
+            .expect("ipc writer");
+        Ok(PyBytes::new_bound(py, &buf).to_object(py))
+    }
+
+    #[cfg(feature = "ipc_streaming")]
+    fn __setstate__(&mut self, py: Python, state: PyObject) -> PyResult<()> {
+        // Used in pickle/pickling
+        match state.extract::<&PyBytes>(py) {
+            Ok(s) => {
+                let c = Cursor::new(s.as_bytes());
+                let reader = IpcStreamReader::new(c);
+
+                reader
+                    .finish()
+                    .map(|df| {
+                        self.df = df;
+                    })
+                    .map_err(|e| PyPolarsErr::from(e).into())
+            },
+            Err(e) => Err(e),
+        }
+    }
+
+    #[cfg(feature = "json")]
+    pub fn serialize(&mut self, py_f: PyObject) -> PyResult<()> {
+        let file = BufWriter::new(get_file_like(py_f, true)?);
+        serde_json::to_writer(file, &self.df)
+            .map_err(|e| polars_err!(ComputeError: "{e}"))
+            .map_err(|e| PyPolarsErr::Other(format!("{e}")).into())
+    }
+
+    #[staticmethod]
+    #[cfg(feature = "json")]
+    pub fn deserialize(py: Python, mut py_f: Bound<PyAny>) -> PyResult<Self> {
+        use crate::file::read_if_bytesio;
+        py_f = read_if_bytesio(py_f);
+        let mut mmap_bytes_r = get_mmap_bytes_reader(&py_f)?;
+
+        py.allow_threads(move || {
+            let mmap_read: ReaderBytes = (&mut mmap_bytes_r).into();
+            let bytes = mmap_read.deref();
+            match serde_json::from_slice::<DataFrame>(bytes) {
+                Ok(df) => Ok(df.into()),
+                Err(e) => {
+                    let msg = format!("{e}");
+                    let e = PyPolarsErr::from(PolarsError::ComputeError(msg.into()));
+                    Err(PyErr::from(e))
+                },
+            }
+        })
+    }
+}
