+++ py-polars/src/expr/mod.rs
@@ -8,6 +8,7 @@ mod list;
 mod meta;
 mod name;
 mod rolling;
+mod serde;
 mod string;
 mod r#struct;
